Refactoring: - Remove unnecessary links to ametsuchi in CMakeLists - Simplify genesis_proc_test - Move AmetsuchiTest fixture to separate header with interface library
