+++ irohad/synchronizer/CMakeLists.txt
@@ -2,7 +2,7 @@ add_library(synchronizer
     impl/synchronizer_impl.cpp
     )
 
-target_link_libraries(synchronizer PUBLIC
+target_link_libraries(synchronizer
     model
     rxcpp
     logger
