+++ test/module/irohad/consensus/yac/CMakeLists.txt
@@ -38,7 +38,6 @@ target_link_libraries(yac_network_test
 addtest(yac_peer_orderer_test peer_orderer_test.cpp)
 target_link_libraries(yac_peer_orderer_test
     yac
-    ametsuchi
     )
 
 addtest(yac_gate_test yac_gate_test.cpp)
