+++ test/module/irohad/ordering/CMakeLists.txt
@@ -15,7 +15,6 @@
 addtest(ordering_service_test ordering_service_test.cpp)
 target_link_libraries(ordering_service_test
     ordering_service
-    ametsuchi
     )
 
 addtest(ordering_gate_test ordering_gate_test.cpp)
@@ -26,5 +25,4 @@ target_link_libraries(ordering_gate_test
 addtest(ordering_gate_service_test ordering_gate_service_test.cpp)
 target_link_libraries(ordering_gate_service_test
     ordering_service
-    ametsuchi
     )
