+++ test/module/irohad/synchronizer/CMakeLists.txt
@@ -1,7 +1,6 @@
 addtest(synchronizer_test synchronizer_test.cpp)
 target_link_libraries(synchronizer_test
-      synchronizer
-      ametsuchi
-      )
+    synchronizer
+    )
