+++ test/module/irohad/torii/CMakeLists.txt
@@ -22,7 +22,6 @@ target_link_libraries(torii_service_test
     query_client
     server_runner
     processors
-    ametsuchi
     )
 
 addtest(torii_queries_test torii_queries_test.cpp)
@@ -32,5 +31,4 @@ target_link_libraries(torii_queries_test
     query_client
     server_runner
     processors
-    ametsuchi
     )
