+++ test/module/irohad/torii/processor/CMakeLists.txt
@@ -8,5 +8,4 @@ target_link_libraries(transaction_processor_test
 addtest(query_processor_test query_processor_test.cpp)
 target_link_libraries(query_processor_test
     processors
-    ametsuchi
     )
