+++ test/module/irohad/validation/CMakeLists.txt
@@ -10,11 +10,9 @@ target_link_libraries(stateless_transaction_validator_test
 addtest(query_execution query_execution.cpp)
 target_link_libraries(query_execution
     model
-    ametsuchi
     )
 
 addtest(chain_validation_test chain_validation_test.cpp)
 target_link_libraries(chain_validation_test
     chain_validator
-    ametsuchi
     )
